Support v2 Stoplight platform
