+++ lib/bundle/stoplight/generator.js
@@ -1,19 +1,36 @@
 "use strict";
 
-const { join, stripRoot } = require("@stoplight/path");
-const {
-  getGenericDefaults,
-  getDefaultsForOldJsonSchema,
-  getDefaultsForOAS2,
-  getDefaultsForOAS3,
-} = require("./defaults");
-const { parse, isHttp, isFileSystemPath } = require("../util/url");
-const KeyGenerator = require("./util/key-generator");
+const { parse, isHttp, isFileSystemPath } = require("../../util/url");
+const KeyGenerator = require("../util/key-generator");
+
+function normalizeOpts (opts) {
+  if (typeof opts.cwd === "string") {
+    opts.cwd = StoplightKeyGenerator.appendSlash(opts.cwd);
+  }
+  else {
+    opts.cwd = null;
+  }
+
+  if (typeof opts.endpointUrl === "string") {
+    opts.endpointUrl = StoplightKeyGenerator.appendSlash(opts.endpointUrl);
+  }
+  else {
+    opts.endpointUrl = null;
+  }
+
+  return opts;
+}
 
 function StoplightKeyGenerator (root, opts) {
   KeyGenerator.call(this, root);
-  this._opts = opts;
+  this._opts = normalizeOpts({ ...opts });
 }
+
+StoplightKeyGenerator.appendSlash = function (str) {
+  return str.replace(/([^/])\/?$/, "$1/");
+};
+
+module.exports = StoplightKeyGenerator;
 
 StoplightKeyGenerator.prototype = Object.create(KeyGenerator.prototype, {
   constructor: {
@@ -28,35 +45,24 @@ StoplightKeyGenerator.prototype.processPath = function (path) {
     return this.extractFilepathFromUrl(path);
   }
 
-  let { srn, cwd } = this._opts;
+  let { cwd } = this._opts;
 
-  if (srn !== null && cwd !== null && isFileSystemPath(path) && path.indexOf(cwd) === 0) {
+  if (path.indexOf(cwd) === 0) {
     // filesystem, file in the same project
-    return join(srn, stripRoot(path.slice(cwd.length)));
+    return path.slice(cwd.length);
   }
 
   return path;
 };
 
 StoplightKeyGenerator.prototype.extractFilepathFromUrl = function (url) {
-  let { endpointUrl, cwd } = this._opts;
+  let { endpointUrl } = this._opts;
   let { href, query } = parse(url, true);
 
   let mid = query.mid ? `_m${query.mid}` : "";
 
-  if (typeof query.srn === "string") {
-    // this is for v1 (srn)
-    return query.srn + mid;
-  }
-
-  if (endpointUrl !== null && href.indexOf(endpointUrl) === 0) {
-    // this is for v1 (href)
+  if (href.indexOf(endpointUrl) === 0) {
     return href.slice(endpointUrl.length) + mid;
-  }
-
-  // this is mostly for the sake of Karma... but might be useful for us, one day. Perhaps.
-  if (cwd !== null && isHttp(cwd) && href.indexOf(cwd) === 0) {
-    return href.slice(cwd.length);
   }
 
   return url;
@@ -111,7 +117,7 @@ StoplightKeyGenerator.prototype.generateKeyForUrl = function (schema, url) {
 
       let generatedKey = this.registerNewGeneratedKey(schema, url, key);
 
-      // https://example.com/api/nodes.raw/?srn=org/proj/data-model-dictionary/reference/book.yaml and "https://example.com/api/nodes.raw/org/proj/data-model-dictionary/reference/book.yaml
+      // https://example.com/api/nodes.raw/?srn=org/proj/data-model-dictionary/reference/book.yaml and filepath from the same project
       // should be considered the same source, hence this extra registration call here
       // query.mid appended because mid does affect the way we treat the source
       this.registerNewGeneratedKey(schema, filepath, key);
@@ -123,19 +129,4 @@ StoplightKeyGenerator.prototype.generateKeyForUrl = function (schema, url) {
   }
 
   return null;
-};
-
-module.exports = function (opts) {
-  return {
-    get oas2 () {
-      return getDefaultsForOAS2(getGenericDefaults(new StoplightKeyGenerator("#/definitions", opts)));
-    },
-    get oas3 () {
-      return getDefaultsForOAS3(getGenericDefaults(new StoplightKeyGenerator("#/components/schemas", opts)));
-    },
-    // eslint-disable-next-line camelcase
-    get json_schema () {
-      return getDefaultsForOldJsonSchema(getGenericDefaults(new StoplightKeyGenerator("#/definitions", opts)));
-    },
-  };
 };
