+++ test/specs/custom-bundling-roots-legacy-stoplight/reference/mocks/airport-masked.json
