+++ test/specs/custom-bundling-roots-legacy-stoplight/reference/openapi-2.json
