+++ test/specs/custom-bundling-roots-legacy-stoplight/reference/openapi-3.json
