+++ test/specs/custom-bundling-roots-legacy-stoplight/reference/parameters/id.json
@@ -1,0 +1,6 @@
+{
+  "type": "number",
+  "name": "id",
+  "in": "path",
+  "required": true
+}
