+++ test/specs/custom-bundling-roots-legacy-stoplight/reference/schemas/alt/user.json
