+++ test/specs/custom-bundling-roots-legacy-stoplight/reference/schemas/common/airplane.v1.json
