+++ test/specs/custom-bundling-roots-legacy-stoplight/reference/schemas/common/manufacturer.json
