+++ test/specs/custom-bundling-roots-legacy-stoplight/reference/schemas/flight.json
