+++ test/specs/custom-bundling-roots-new-stoplight/reference/Book.v1.json
@@ -1,0 +1,3 @@
+{
+  "title": "Book"
+}
