+++ test/specs/custom-bundling-roots/id.json
